+++ CHANGES.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.2.0] - 2020-04-04
+### Added
+* Finish date calculations now calculates if a task cannot finish in time.
 ### Fixed
 * Finished tasks are no longer marked as having an uncertain finish date (fixes [#2](https://github.com/derkork/project.txt-js/issues/2)).